Create test project for granchild object access
This test project demonstates what happens with two layers of
inheritance in entities.

The `SetByDerivedCircle` object on BaseEntity is expected to be visible
and configurable on GrandchildEntity since it should be settable at that
level. Variables with Access/SetByDerived set properly appear at each
level of inheritance, as do objects with Access/ExposedInDerived set,
but objects with Access/SetByDerived set only appear for the immediately
derived entity.